feat: validações feitas e metodos do controller de usuario finalizados.

+++ app/Controllers/Http/PerfisController.ts
@@ -1,5 +1,7 @@
 import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import Perfil from 'App/Models/Perfil'
+import CreatePerfilValidator from 'App/Validators/Perfil/CreatePerfilValidator'
+import UpdatePerfilValidator from 'App/Validators/Perfil/UpdatePerfilValidator'
 
 export default class PerfisController {
   public async index({ request }: HttpContextContract) {
@@ -10,6 +12,8 @@ export default class PerfisController {
 
   public async store({ request }: HttpContextContract) {
     const data = request.only(['nome'])
+
+    await request.validate(CreatePerfilValidator)
 
     const perfil = await Perfil.create(data)
 
@@ -24,22 +28,32 @@ export default class PerfisController {
     return perfil
   }
 
-  public async update({ request, params }: HttpContextContract) {
+  public async update({ request, params, response }: HttpContextContract) {
     const { id } = params;
 
     const data = request.only(['nome'])
 
-    const perfil = await Perfil.findOrFail(id)
+    const perfil = await Perfil.find(id)
 
-    perfil.merge(data)
+    if (!perfil) {
+      return response.notFound({ message: 'Perfil não encontrado' })
+    }
+
+    await request.validate(UpdatePerfilValidator)
+  
+    await perfil.merge(data).save()
 
     return perfil
   }
 
-  public async destroy({ params }: HttpContextContract) {
+  public async destroy({ params, response }: HttpContextContract) {
     const { id } = params;
 
-    const perfil = await Perfil.findOrFail(id)
+    const perfil = await Perfil.find(id)
+
+    if (!perfil) {
+      return response.notFound({ message: 'Perfil não encontrado' })
+    }
 
     perfil.delete()
